extract otel connector
