+++ tremor-connectors-gcp/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2021"
 [dependencies]
 tremor-connectors = { path = "../tremor-connectors", version = "0.13.0-rc.20", default-features = false, features = [
     "object-storage",
-    "protobuf",
 ] }
 tremor-common = { path = "../tremor-common", version = "0.13.0-rc.20" }
 tremor-system = { path = "../tremor-system", version = "0.13.0-rc.20" }
@@ -75,10 +74,8 @@ tempfile = { version = "3", default-features = true }
 env_logger = { version = "0.11", default-features = true }
 tremor-connectors = { path = "../tremor-connectors", default-features = false, features = [
     "object-storage",
-    "protobuf",
     "integration-utils",
 ] }
-
 
 [features]
 integration-harness = ["tremor-connectors/integration-utils"]
