+++ tremor-connectors-otel/Cargo.toml
@@ -1,0 +1,47 @@
+[package]
+name = "tremor-connectors-otel"
+version = "0.13.0-rc.20"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+tremor-connectors = { path = "../tremor-connectors", version = "0.13.0-rc.20", default-features = false }
+tremor-common = { path = "../tremor-common", version = "0.13.0-rc.20" }
+tremor-system = { path = "../tremor-system", version = "0.13.0-rc.20" }
+tremor-value = { path = "../tremor-value", version = "0.13.0-rc.20" }
+tremor-config = { path = "../tremor-config", version = "0.13.0-rc.20" }
+tremor-script = { path = "../tremor-script", version = "0.13.0-rc.20" }
+
+anyhow = { version = "1", default-features = true }
+async-trait = { version = "0.1", default-features = true }
+futures = { version = "0.3", default-features = true }
+http = { version = "0.2", default-features = true }
+http-body = { version = "1.0", default-features = true }
+log = { version = "0.4", default-features = true }
+serde = { version = "1.0", default-features = true, features = ["derive"] }
+thiserror = { version = "1", default-features = true }
+value-trait = { version = "0.8", default-features = true }
+tokio = { version = "1.34", default-features = false }
+simd-json = { version = "0.13", default-features = true }
+
+# opentelemetry
+tonic = { version = "0.6.1", default-features = false, features = [
+    "transport",
+    "tls",
+] }
+prost = { version = "0.9.0", default-features = false }
+prost-types = { version = "0.9.0", default-features = false }
+# This is related to https://github.com/tremor-rs/tremor-runtime/issues/1688 the otel API's need to
+# be updated together with tonic
+tremor-otelapis = { version = "=0.2.4", default-features = false, features = [
+    "otel-all",
+] }
+rand = { version = "0.8.5", default-features = false }
+async-channel = { version = "1.9", default-features = false }
+hex = { version = "0.4", default-features = false }
+
+[dev-dependencies]
+tokio = { version = "1.34", default-features = true }
+proptest = "1.4"
+test-case = "3"
